// backend/src/domain/enums/dependency.enums.ts
export enum DependencyUrgency {
  Critical = 'critical',
  Recommended = 'recommended',
  Optional = 'optional',
}

export enum DependencyEcosystem {
  Npm = 'npm',
  Pip = 'pip',
  Composer = 'composer',
  Cargo = 'cargo',
  Maven = 'maven',
  Gradle = 'gradle',
  Go = 'go',
  Nuget = 'nuget',
  RubyGems = 'rubygems',
}